fix(cl): Update code for PR requests
Various fixes and tweaks to reflect @mlissner
comments on the current PR
